Move NoteWidget to a separate file
